refactor:  Application 클래스의 게임 실행 로직을 BridgeGameController 에게 위임
